+++ forge/src/main/java/de/maxhenkel/voicechat/ForgeVoicechatClientMod.java
@@ -4,6 +4,7 @@ import com.sun.jna.Platform;
 import de.maxhenkel.voicechat.config.ForgeClientConfig;
 import de.maxhenkel.voicechat.gui.VoiceChatSettingsScreen;
 import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
+import de.maxhenkel.voicechat.intercompatibility.ForgeClientCompatibilityManager;
 import net.minecraftforge.client.ConfigGuiHandler;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.ModLoadingContext;
@@ -15,6 +16,7 @@ public class ForgeVoicechatClientMod extends VoicechatClient {
 
     public ForgeVoicechatClientMod() {
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::clientSetup);
+        FMLJavaModLoadingContext.get().getModEventBus().addListener(((ForgeClientCompatibilityManager) ClientCompatibilityManager.INSTANCE)::onRegisterKeyBinds);
         CLIENT_CONFIG = ForgeVoicechatMod.registerConfig(ModConfig.Type.CLIENT, ForgeClientConfig::new);
     }